Move grabcut to smart-tools; Update worker
